Split OpenMPIR.cpp to three files
Refactor ompparser OpenMPIR.cpp code into three files: OpenMPIR.cpp for IR construction and normalization, OpenMPIRToString.cpp, OpenMPIRToDOT.cpp files for unparsing OpenMPIR to a string and a dot file.
