+++ CMakeLists.txt
@@ -19,6 +19,14 @@ set(OMPPARSER_SOURCE_FILES
     src/omplexer.ll
     src/ompparser.yy
     src/OpenMPIR.h
+    src/OpenMPIRToDOT.cpp
+    src/OpenMPIRToString.cpp
+    src/OpenMPIR.cpp)
+
+# OpenMPIR source files
+set(OMPIR_SOURCE_FILES
+    src/OpenMPIRToDOT.cpp
+    src/OpenMPIRToString.cpp
     src/OpenMPIR.cpp)
 
 BISON_TARGET(OMPBISONParser src/ompparser.yy ${CMAKE_CURRENT_BINARY_DIR}/ompparser.yy.cpp)
@@ -29,13 +37,13 @@ include_directories(${CMAKE_CURRENT_BINARY_DIR})
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/src)
 
 # The temporary test file
-add_executable(main ${BISON_OMPBISONParser_OUTPUTS} ${FLEX_OMPFLEXScanner_OUTPUTS} src/OpenMPIR.cpp main.cpp)
+add_executable(main ${BISON_OMPBISONParser_OUTPUTS} ${FLEX_OMPFLEXScanner_OUTPUTS} ${OMPIR_SOURCE_FILES} main.cpp)
 
 # The temporary test file
-add_executable(tester ${BISON_OMPBISONParser_OUTPUTS} ${FLEX_OMPFLEXScanner_OUTPUTS} src/OpenMPIR.cpp tester.cpp)
+add_executable(tester ${BISON_OMPBISONParser_OUTPUTS} ${FLEX_OMPFLEXScanner_OUTPUTS} ${OMPIR_SOURCE_FILES} tester.cpp)
 
 # The ompparser library for install and distribution
-add_library(ompparser SHARED ${BISON_OMPBISONParser_OUTPUTS} ${FLEX_OMPFLEXScanner_OUTPUTS} src/OpenMPIR.cpp)
+add_library(ompparser SHARED ${BISON_OMPBISONParser_OUTPUTS} ${FLEX_OMPFLEXScanner_OUTPUTS} ${OMPIR_SOURCE_FILES})
 
 set(ompparser_targets ompparser)
